% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myEigen_QR.R
\name{myEigen_QR}
\alias{myEigen_QR}
\title{Perform PCA on matrix A using QR}
\usage{
myEigen_QR(A, numIter = 1000)
}
\arguments{
\item{matrix}{A is an square matrix}

\item{int}{numIter is the number of iterations}
}
\value{
a list with D and V, where D is a vector of eigenvalues of A 
        and V is the matrix of eigenvectors of A
}
\description{
Perform PCA on matrix A using QR
}
