% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLasso.R
\name{myLasso}
\alias{myLasso}
\title{Find the lasso solution path for various values of 
the regularization parameter lambda.}
\usage{
myLasso(X, Y, lambda_all)
}
\arguments{
\item{matrix}{n x p matrix of explanatory variables}

\item{matrix}{n dimensional response vector}

\item{vector}{Vector of regularization parameters}
}
\value{
a matrix containing the lasso solution vector 
  beta for each regularization parameter.
}
\description{
Find the lasso solution path for various values of 
the regularization parameter lambda.
}
