% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLogistic.R
\name{myLogistic}
\alias{myLogistic}
\title{Perform the logistic regression of Y on X}
\usage{
myLogistic(X, Y)
}
\arguments{
\item{matrix}{X is an n x p matrix of explanatory variables}

\item{matrix}{Y is an n dimensional vector of binary responses}

\item{int}{numIter is the number of iterations}
}
\value{
a list where the first element is a 1 x p vector beta and the second
        element is the standard error
}
\description{
Perform the logistic regression of Y on X
}
