% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myQR.R
\name{myQR}
\alias{myQR}
\title{Perform QR decomposition on matrix A}
\usage{
myQR(A)
}
\arguments{
\item{matrix}{input n x m matrix A}
}
\value{
a list with Q.transpose and R, where Q is an orthogonal n x n matrix and
        R is an upper triangular n x m matrix
}
\description{
Perform QR decomposition on matrix A
}
