% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myRidge.R
\name{myRidge}
\alias{myRidge}
\title{Perform ridge regression of Y on X.}
\usage{
myRidge(X, Y, lambda)
}
\arguments{
\item{matrix}{n x p matrix of explanatory variables}

\item{matrix}{n dimensional response vector}

\item{vector}{regularization parameters}
}
\value{
Returns beta, the ridge regression solution.
}
\description{
Perform ridge regression of Y on X.
}
